feat/MSSDK-2048: Changes after CR
