+++ .env
@@ -1,1 +1,2 @@
 VITE_MEDIASTORE_SDK_VERSION=$npm_package_version
+VITE_RECAPTCHA_SITE_KEY="6Ld0A54qAAAAANJ8mLCpJAxEp0XKtJyueFmEFVaG"
