+++ src/types/global.types.ts
@@ -1,0 +1,11 @@
+export {};
+
+type RecaptchaOptions = {
+  enterprise: boolean;
+};
+
+declare global {
+  interface Window {
+    recaptchaOptions: RecaptchaOptions;
+  }
+}
